+++ examples/chat/src/components/mod.rs
@@ -1,0 +1,7 @@
+mod chat;
+mod message;
+mod rooms;
+
+pub use chat::*;
+pub use message::*;
+pub use rooms::*;